+++ waffle_hub/hub/base_hub.py
@@ -11,11 +11,14 @@ from functools import cached_property
 from pathlib import Path
 from typing import Union
 
+import cv2
 import torch
+import tqdm
 from waffle_utils.file import io
 from waffle_utils.utils import type_validator
 
 from waffle_hub.schemas.configs import Model, Train
+from waffle_hub.utils.image import draw_results
 
 logger = logging.getLogger(__name__)
 
@@ -476,8 +479,44 @@ class BaseHub:
     def on_inference_start(self, ctx: InferenceContext):
         pass
 
-    def inferencing(self, ctx: InferenceContext):
-        pass
+    def inferencing(self, ctx: InferenceContext) -> str:
+        model = ctx.model.to(ctx.device)
+        dataloader = ctx.dataloader
+        device = ctx.device
+
+        for images, image_infos in tqdm.tqdm(dataloader):
+            result_batch = model(images.to(device), image_infos)
+            for results, image_info in zip(result_batch, image_infos):
+                image_path = image_info.get("image_path")
+
+                relpath = Path(image_path).relative_to(ctx.source)
+                if relpath == Path("."):
+                    relpath = Path(
+                        "test.png"
+                    )  # TODO: path correction for none directory source.
+                io.save_json(
+                    results,
+                    self.inference_dir
+                    / "results"
+                    / relpath.with_suffix(".json"),
+                    create_directory=True,
+                )
+                if ctx.draw:
+                    draw = draw_results(
+                        image_path,
+                        results,
+                        task=self.task,
+                        names=[x["name"] for x in self.classes],
+                    )
+                    draw_path = (
+                        self.inference_dir
+                        / "draw"
+                        / relpath.with_suffix(".png")
+                    )
+                    io.make_directory(draw_path.parent)
+                    cv2.imwrite(str(draw_path), draw)
+
+        return results
 
     def on_inference_end(self, ctx: InferenceContext):
         pass
